Extract RBS comments visit as a mixin
