+++ lib/spoom.rb
@@ -17,6 +17,7 @@ require "spoom/poset"
 require "spoom/model"
 require "spoom/source"
 require "spoom/deadcode"
+require "spoom/rbs"
 require "spoom/sorbet"
 require "spoom/cli"
 require "spoom/version"
